backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.9.0
firebase-admin>=6.5.0
google-cloud-storage>=2.17.0
websockets>=13.0
google-genai
google-auth
